requirements.txt
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
oauth2client==4.1.3
python-dotenv==1.0.0
requests==2.28.2


